Replace raw setInterval call with interval web worker in IntervalClock.
Because the worker runs on a separate thread this should make the clock resilient to expensive things on the main thread like resizing the window or calculating prime numbers. However, once we are ready to sync the metronome animation to audio then we'll probably stop using clocks altogether and just use requestAnimationFrame and query the audio context.